+++ peers/index.js
@@ -1,5 +1,6 @@
+const connectPeer = require('./connect_peer');
 const findKey = require('./find_key');
 const getLiquidity = require('./get_liquidity');
 const getPeerLiquidity =  require('./get_peer_liquidity');
 
-module.exports = {findKey, getLiquidity, getPeerLiquidity};
+module.exports = {connectPeer, findKey, getLiquidity, getPeerLiquidity};